+++ src/theme/components/table.ts
@@ -56,14 +56,19 @@ export const fixedColumnStyles = (
   },
 });
 
-const tableTextSizes = (textSizeValue: string = "body1") => ({
-  columnHeading:
-    textSizeValue === "body2"
-      ? "desktop.caption.caption1"
-      : "desktop.body.body2",
-  tableBody:
-    textSizeValue === "body2" ? "desktop.body.body2" : "desktop.body.body1",
-});
+const tableTextSizes = (textSizeValue: string = "body1") => {
+  const fontSizeMap = {
+    body1: "desktop.body.body1",
+    body2: "desktop.body.body2",
+    caption: "desktop.caption",
+  };
+
+  return {
+    columnHeading:
+      textSizeValue === "body1" ? "desktop.body.body2" : "desktop.caption",
+    tableBody: fontSizeMap[textSizeValue],
+  };
+};
 
 export const baseTRStyles = (
   columnHeadersBackgroundColor = "",